Remove Gemfile.lock
All the gems (except appmap) are pinned to exact versions, so removing
the lockfile isn't going to cause problems. Doing this to workaround the
following error when trying to update appmap:

```
Unable to find a spec satisfying appmap (>= 0) in the set. Perhaps the lockfile
is corrupted?
```
As far as I can tell, the lockfile *wasn't* corrupted though, so I'm
not entirely sure what's going on.

Also fixes the syntax of .travis.yml.